Added driver changes

+++ src/main/java/edu/neu/csye7374/strategypattern/OnlinePaymentStrategy.java
@@ -14,7 +14,7 @@ public class OnlinePaymentStrategy implements PaymentStrategy{
     }
 
     @Override
-    public void processPayment(int totalBillAmount) {
+    public void processPayment(double totalBillAmount) {
         System.out.println("CASH PAYMENT DETAILS");
         System.out.println("Total Bill Amount " + totalBillAmount);
         System.out.println("Tip Amount " + this.tipAmount);
